refactor: rename models and update interfaces for clarity and consistency
