+++ src/app/(tabs)/progress.tsx
@@ -5,7 +5,7 @@ import { Feather } from "@expo/vector-icons";
 import { LineChart } from "react-native-gifted-charts";
 import { useAuth, useUser } from "@clerk/clerk-expo";
 import { useFocusEffect } from "@react-navigation/native";
-import { getRatingsByUserId } from "@/api";
+import { getProgress } from "@/api/ratings";
 
 const Progress = () => {
   const { user } = useUser();
@@ -32,7 +32,7 @@ const Progress = () => {
     try {
       const token = await getToken();
       const weekStartFormatted = currentWeekStart.format("YYYY-MM-DD");
-      const fetchedProgress = await getRatingsByUserId(
+      const fetchedProgress = await getProgress(
         weekStartFormatted,
         token
       );
