+++ src/types/FeedbackData.ts
@@ -1,9 +1,17 @@
-export interface FeedbackData {
+export interface FeedbackDataRecord {
     answer_id: string;
     interview_id: string | string[];
     answer: string;
     question: string;
-    wpm: number;
+    pace_of_speech: number;
     eye_contact: number;
+  }
+  
+  export interface FeedbackDataVirtual {
+    interview_id: string | string [];
+    answers: string[];
+    questions: string[];
+    wpm: number[];
+    eye_contact: number[];
   }
