+++ src/types/QuestionData.ts
@@ -1,4 +1,10 @@
 export interface QuestionData {
-  interview_id: string | string[];
-  question: string;
+  type: string;
+  industry: string;
+  experience_level: string;
+  interview_type: string;
+  job_description: string | null;
+  company_name: string | null;
+  job_role: string;
+  interview_id: string;
 }
