+++ src/types/stepContent.ts
@@ -6,7 +6,5 @@ export interface StepContentProps {
   updateFormData: (key: string, value: string, callback?: () => void) => void;
   handleNextStep: () => void;
   handleSubmit: () => void;
-  handleSubmitRoute: string;
   handleSkip: () => void;
-  interviewId: string | null;
 }
